[FEAT] Show Kingdom Chores in Faction Tab in Codex
