+++ src/features/island/hud/components/codex/pages/FactionLeaderboard.tsx
@@ -28,6 +28,7 @@ import { formatNumber, shortenCount } from "lib/utils/formatNumber";
 import { NPCName, NPC_WEARABLES } from "lib/npcs";
 import { ChampionsPrizes } from "features/world/ui/factions/Champions";
 import { toOrdinalSuffix } from "features/retreat/components/auctioneer/AuctionLeaderboardTable";
+import { KingdomChores } from "features/world/ui/factions/chores/KingdomChoresCodex";
 
 const npcs: Record<FactionName, NPCName> = {
   nightshades: "nyx",
@@ -165,6 +166,7 @@ export const FactionLeaderboard: React.FC<Props> = ({
           </div>
         </div>
         <ChampionsPrizes />
+        <KingdomChores />
       </div>
     </InnerPanel>
   );
